api/requirements.txt


tensorflow==2.13.*
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.13.*
matplotlib
numpy
pydantic